feat: refatoração e comentarios
