+++ my-app/src/pages/AdminLoginPage.css
@@ -58,6 +58,7 @@
   }
   
   .adm-signup-link {
+    color: #7f00ff;
     margin-top: 20px;
     font-size: 0.9rem;
   }
