+++ my-app/src/pages/LoginPage.css
@@ -59,6 +59,7 @@
   
   .signup-link {
     margin-top: 20px;
+    color:  #0072ff;
     font-size: 0.9rem;
   }
